Move joining rooms tests to their own class
